refactor: Fix some folders and imports
