+++ src/components/layout/Footer.tsx
@@ -1,5 +1,5 @@
 import { Flex } from "@chakra-ui/react";
-import { HEIGHT } from "@/constants.ts/ui";
+import { HEIGHT } from "@/constants/ui";
 
 const Footer = () => (
   <Flex as="footer" h={`${HEIGHT.footer}px`}>
